wip Inflation

+++ src/Inflation/Inflation.sol
@@ -20,20 +20,20 @@ contract Inflation is Ownable {
     constructor(IERC20 iQ, Reward[] memory rewards) public Ownable() {
         _iQ = iQ;
         _lastClaimed = block.timestamp;
-        uint rewardsLength = rewards.length;
-        for (uint i = 0; i < rewardsLength; i++) {
+        uint256 rewardsLength = rewards.length;
+        for (uint256 i = 0; i < rewardsLength; i++) {
             _rewards.push(rewards[i]);
         }
     }
 
     function inflate() external {
-        uint rewardsLength = _rewards.length;
-        uint currentTimestamp = block.timestamp;
-        uint diffSeconds = currentTimestamp - _lastClaimed;
+        uint256 rewardsLength = _rewards.length;
+        uint256 currentTimestamp = block.timestamp;
+        uint256 diffSeconds = currentTimestamp - _lastClaimed;
         require(diffSeconds > 600, "maximum one call each 10 minutes");
-        for (uint i = 0; i < rewardsLength; i++) {
+        for (uint256 i = 0; i < rewardsLength; i++) {
             Reward memory reward = _rewards[i];
-            uint emission = reward.emissionsPerSecond * diffSeconds;
+            uint256 emission = reward.emissionsPerSecond * diffSeconds;
             require(_iQ.transfer(reward.destination, emission), "Transfer has failed");
             emit Inflate(reward.destination, emission);
         }
@@ -42,7 +42,7 @@ contract Inflation is Ownable {
 
     function changeInflation(Reward[] memory rewards) external onlyOwner {
         delete _rewards;
-        for (uint i = 0; i < rewards.length; i++) {
+        for (uint256 i = 0; i < rewards.length; i++) {
             _rewards.push(rewards[i]);
         }
     }
